ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
transformers==4.41.2
torch==2.3.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
pydantic==2.7.1
python-dotenv==1.0.1